+++ tests/test_hookman_generator/hook_specs_no_pyd.py
@@ -1,0 +1,7 @@
+from hookman.hooks import HookSpecs
+
+specs = HookSpecs(
+    project_name='ACME',
+    version='1',
+    hooks=[],
+)